+++ comparable_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/fluentassert/verify"
 )
 
-func TestComparable(t *testing.T) {
+func TestObj(t *testing.T) {
 	type A struct {
 		Str  string
 		Bool bool
@@ -16,18 +16,18 @@ func TestComparable(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true}
 			got := A{Str: "string", Bool: true}
-			msg := verify.Comparable(got).Equal(want)
+			msg := verify.Obj(got).Equal(want)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true}
 			got := A{Str: "wrong", Bool: true}
-			msg := verify.Comparable(got).Equal(want)
+			msg := verify.Obj(got).Equal(want)
 			assertFailed(t, msg, "the objects are not equal")
 		})
 		t.Run("nil", func(t *testing.T) {
 			var got *A
-			msg := verify.Comparable(got).Equal(nil)
+			msg := verify.Obj(got).Equal(nil)
 			assertPassed(t, msg)
 		})
 	})
@@ -36,18 +36,18 @@ func TestComparable(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true}
 			got := A{Str: "wrong", Bool: true}
-			msg := verify.Comparable(got).NotEqual(want)
+			msg := verify.Obj(got).NotEqual(want)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true}
 			got := A{Str: "string", Bool: true}
-			msg := verify.Comparable(got).NotEqual(want)
+			msg := verify.Obj(got).NotEqual(want)
 			assertFailed(t, msg, "the objects are equal")
 		})
 		t.Run("nil", func(t *testing.T) {
 			var got *A
-			msg := verify.Comparable(got).NotEqual(nil)
+			msg := verify.Obj(got).NotEqual(nil)
 			assertFailed(t, msg, "the objects are equal")
 		})
 	})
@@ -55,17 +55,17 @@ func TestComparable(t *testing.T) {
 	t.Run("Zero", func(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
 			got := A{}
-			msg := verify.Comparable(got).Zero()
+			msg := verify.Obj(got).Zero()
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			got := A{Str: "wrong"}
-			msg := verify.Comparable(got).Zero()
+			msg := verify.Obj(got).Zero()
 			assertFailed(t, msg, "not a zero value\n")
 		})
 		t.Run("nil", func(t *testing.T) {
 			var got *A
-			msg := verify.Comparable(got).Zero()
+			msg := verify.Obj(got).Zero()
 			assertPassed(t, msg)
 		})
 	})
@@ -73,24 +73,24 @@ func TestComparable(t *testing.T) {
 	t.Run("NonZero", func(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
 			got := A{Str: "string"}
-			msg := verify.Comparable(got).NonZero()
+			msg := verify.Obj(got).NonZero()
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			got := A{}
-			msg := verify.Comparable(got).NonZero()
+			msg := verify.Obj(got).NonZero()
 			assertFailed(t, msg, "a zero value")
 		})
 		t.Run("nil", func(t *testing.T) {
 			var got *A
-			msg := verify.Comparable(got).NonZero()
+			msg := verify.Obj(got).NonZero()
 			assertFailed(t, msg, "a zero value")
 		})
 	})
 
 	t.Run("has assertions from Any", func(t *testing.T) {
 		want := A{}
-		got := verify.Comparable(want).FluentAny.Got // type embedding done properly
+		got := verify.Obj(want).FluentAny.Got // type embedding done properly
 		assertEqual(t, got, want)
 	})
 }