+++ string_test.go
@@ -140,9 +140,9 @@ func TestString(t *testing.T) {
 		})
 	})
 
-	t.Run("has assertions from Ordered, Comparable, Obj", func(t *testing.T) {
+	t.Run("has assertions from Ordered, Obj, Any", func(t *testing.T) {
 		want := "text"
-		got := verify.String(want).FluentOrdered.FluentComparable.FluentAny.Got // type embedding done properly
+		got := verify.String(want).FluentOrdered.FluentObj.FluentAny.Got // type embedding done properly
 		assertEqual(t, got, want)
 	})
 }